+++ CHANGELOG.md
@@ -12,6 +12,9 @@
     - Refactored `ColorGlyphs` and `ColorPalettes` to parse on demand.
     - Windows builds with default fonts now uses ~20MB less memory.
 
+* Fix gradient stops with midway adjustment.
+* Replace `layout::Vector` add and sub implementations to be generic over any type that converts to vector.
+* Add `Var::chase_begin` to begin a deferred chase animation.
 * Add `zng::app::memory_profiler` for recording DHAT heap traces.
 * Add `zng::task::set_spawn_panic_handler` for apps to optionally handle panics in spawn and forget tasks.
 * Task worker process timeout is now configurable with ZNG_TASK_WORKER_TIMEOUT env var.